backend/requirements.txt
# FastAPI and server
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
websockets>=12.0

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1
asyncpg>=0.29.0

# Authentication & Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0

# HTTP requests
httpx>=0.25.2
requests>=2.31.0

# File processing
python-docx>=1.1.0
PyPDF2>=3.0.1
pandas>=2.1.4
aiofiles>=23.2.1

# Vector database and AI
pinecone>=5.0.0
openai>=1.6.1

# AI Models
groq>=0.4.1
anthropic>=0.8.0

# LangChain
langchain
langchain-openai
langchain-groq

# Validation
pydantic>=2.11.0
pydantic-settings>=2.9.0

# Stripe payments
stripe>=7.0.0

# Google OAuth
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0

# Additional utilities
email-validator>=2.2.0
